Move helper functions into discordtools and timetools
