+++ main.py
@@ -4,6 +4,8 @@ import discord
 from discord.ext import commands, tasks
 import discord.ext.commands
 from bot.tools import *
+from bot.timetools import *
+from bot.discordtools import *
 import firebase_admin
 from firebase_admin import credentials
 from firebase_admin import db
@@ -132,11 +134,10 @@ async def addtask(ctx, *, args):
 
         try:
             time_delta = parse_time_delta(time)
-            due_time = datetime.datetime.now().replace(second=0, microsecond=0) + time_delta
-            unix_timestamp = int(due_time.timestamp())
+            unix_timestamp = duedate_in_unix_timestamp(time_delta)
             
             # Convert ctx.author.id to a string and remove any invalid characters
-            author_id = str(ctx.author.id).replace('.', '_').replace('$', '_').replace('#', '_').replace('[', '_').replace(']', '_')
+            author_id = discord_id_to_firebase_entry(ctx.author.id)
 
             ref = db.reference(f'users/{author_id}/tasks')
             if not ref.get():
@@ -180,33 +181,6 @@ async def addtask_error(ctx, error):
         embed.set_footer(text=ctx.author.display_name, icon_url=ctx.author.avatar)
         await ctx.send(embed=embed)
 
-def parse_time_delta(time_string):
-    # Initialize a total timedelta
-    total_delta = datetime.timedelta()
-    
-    # Regular expression to match time components
-    pattern = r'(\d+)\s*(day|days|week|weeks|h|hour|hours|minute|minutes|min)'
-    
-    # Find all matches in the input string
-    matches = re.findall(pattern, time_string, re.IGNORECASE)
-
-    if not matches:
-        raise commands.BadArgument("Invalid time format. Please use formats like '2 days', '1 hour and 3 minutes', etc.")
-    
-    for amount, unit in matches:
-        amount = int(amount)
-        unit = unit.lower()
-        
-        if unit in ['day', 'days']:
-            total_delta += datetime.timedelta(days=amount)
-        elif unit in ['week', 'weeks']:
-            total_delta += datetime.timedelta(weeks=amount)
-        elif unit in ['hour', 'hours', 'h']:
-            total_delta += datetime.timedelta(hours=amount)
-        elif unit in ['minute', 'minutes', 'min']:
-            total_delta += datetime.timedelta(minutes=amount)
-
-    return total_delta
 
 @bot.command(name='8ball', help='Consults the magic 8-ball :8_ball: !8ball <Question>')
 async def eball(ctx, args):
@@ -241,10 +215,7 @@ async def view(ctx):
     tasks = ref.get()
     if tasks:
         # task[0] = id, task[1] = the items -- in tasks.items()
-        tasks = dict(sorted(tasks.items(), key=lambda x: x[1]['time']))
-        body = ""
-        for idx, (task_id, task) in enumerate(tasks.items()):
-            body += f"{idx + 1}: '{task['task']}' by <t:{int(task['time'])}>\n"
+        body = stringify_tasks(tasks)
         embed = discord.Embed(title="Reminders:",
                             description=body,
                             color=discord.Color.green())
@@ -272,7 +243,7 @@ async def delete(ctx, args):
         tasks = ref.get()
         if tasks:
             # task[0] = id, task[1] = the items -- in tasks.items()
-            tasks = sorted(tasks.items(), key=lambda x: x[1]['time'])
+            tasks = tasks_sorted_by_time(tasks)
             if index > len(tasks):
                 embed = discord.Embed(title="Error!",
                                     description=f"Can't delete {index}th task if you only have {len(tasks)} tasks!",
@@ -321,7 +292,7 @@ async def changedate(ctx, *, args):
         tasks = ref.get()
         if tasks:
             # task[0] = id, task[1] = the items -- in tasks.items()
-            tasks = sorted(tasks.items(), key=lambda x: x[1]['time'])
+            tasks = tasks_sorted_by_time(tasks)
             if index > len(tasks):
                 embed = discord.Embed(title="Error!",
                                     description=f"Can't delete {index}th task if you only have {len(tasks)} tasks!",
